+++ src/main/java/racingcar/domain/CarRandomNumberGenerator.java
@@ -1,7 +1,5 @@
 package racingcar.domain;
 
-import racingcar.RandomNumberGenerator;
-
 public class CarRandomNumberGenerator implements RandomNumberGenerator {
 
     private static final int BOUND = 10;
