+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,7 +1,5 @@
 package racingcar.domain;
 
-import racingcar.RandomNumberGenerator;
-
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
