+++ src/test/java/racingcar/domain/CarsTest.java
@@ -3,7 +3,6 @@ package racingcar.domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.RandomNumberGenerator;
 
 import java.util.List;
 import java.util.stream.Collectors;
